Extract plugin-api-execution module
Move the ten simulation and tracing contracts into ethereum/core/plugin-api-execution, published as besu-plugin-api-execution and re-exported by plugin-api. Packages untouched. Relocate the adapters to ethereum/core and ethereum/api and fix javadoc on the moved contracts.

Closes #11145.
