+++ app/src/main/java/org/hyperledger/besu/services/BesuPluginServiceRegistrar.java
@@ -17,10 +17,12 @@ package org.hyperledger.besu.services;
 import org.hyperledger.besu.Runner;
 import org.hyperledger.besu.controller.BesuController;
 import org.hyperledger.besu.ethereum.api.pluginadapter.HealthCheckServiceImpl;
+import org.hyperledger.besu.ethereum.api.pluginadapter.TraceServiceImpl;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
 import org.hyperledger.besu.ethereum.chain.pluginadapter.RlpConverterServiceImpl;
 import org.hyperledger.besu.ethereum.core.MiningConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.pluginadapter.TransactionPoolServiceImpl;
+import org.hyperledger.besu.ethereum.transaction.pluginadapter.BlockSimulatorServiceImpl;
 import org.hyperledger.besu.ethereum.worldstate.pluginadapter.WorldStateServiceImpl;
 import org.hyperledger.besu.plugin.services.BesuEvents;
 import org.hyperledger.besu.plugin.services.BlockSimulationService;
