+++ ethereum/core/plugin-api-execution/build.gradle
@@ -1,0 +1,58 @@
+/*
+ * Copyright contributors to Besu.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+// override the path-derived name (besu-ethereum-core-plugin-api-execution) with the plugin API family naming
+def executionApiArtifactId = 'besu-plugin-api-execution'
+
+jar {
+  archiveBaseName = executionApiArtifactId
+  manifest {
+    attributes(
+      'Specification-Title': archiveBaseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': archiveBaseName,
+      'Implementation-Version': calculateVersion(),
+      'Commit-Hash': getGitCommitDetails(40).hash,
+      'Automatic-Module-Name': 'org.hyperledger.besu.plugin.api.execution'
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api:core')
+  // Address, CallParameter, Hash, StateOverrideMap, Transaction and Wei appear in this module's own signatures
+  api project(':datatypes')
+  // OperationTracer, WorldView and WorldUpdater appear in the contract signatures
+  api project(':evm')
+  // PluginBlockSimulationResult holds an Optional<TrieLog>
+  api project(':ethereum:core:plugin-api-worldstate')
+  // Bytes and Bytes32 appear in BlockOverrides and TransactionSimulationResult
+  api 'io.consensys.tuweni:tuweni-bytes'
+}
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) {
+      groupId = 'org.hyperledger.besu'
+      artifactId = executionApiArtifactId
+      pom {
+        name = 'Besu Plugin API Execution'
+        description = 'Simulation and tracing contracts of the Besu plugin API, letting plugins simulate transactions and blocks, replay and trace blocks, and supply tracers for block import'
+      }
+    }
+  }
+}
