+++ platform/build.gradle
@@ -60,6 +60,7 @@ dependencies {
     api project(':ethereum:blockcreation')
     api project(':ethereum:core')
     api project(':ethereum:core:plugin-api-chain')
+    api project(':ethereum:core:plugin-api-execution')
     api project(':ethereum:core:plugin-api-worldstate')
     api project(':ethereum:core:plugin-api-worldstate-backend')
     api project(':ethereum:eth')
