+++ settings.gradle
@@ -44,6 +44,7 @@ include 'ethereum:api:plugin-api'
 include 'ethereum:blockcreation'
 include 'ethereum:core'
 include 'ethereum:core:plugin-api-chain'
+include 'ethereum:core:plugin-api-execution'
 include 'ethereum:core:plugin-api-worldstate'
 include 'ethereum:core:plugin-api-worldstate-backend'
 include 'ethereum:eth'
